# Optimize fore, back and mode of Colorize::Object(T)