@z64 2018-10-12
I can't reproduce this benchmark, they are exactly the same performance.
```
new fore 247.75M (  4.04ns) (_ 2.96%)  0 B/op   1.00_ slower
old fore 248.04M (  4.03ns) (_ 2.89%)  0 B/op        fastest
```

I get almost exactly your performance if I do not run with `--release` (necessary to enable optimizations):
```
new fore  21.46M ( 46.61ns) (_ 2.33%)  0 B/op        fastest
old fore  15.99M ( 62.56ns) (_ 7.34%)  0 B/op   1.34_ slower
```

Did you do this? There should be a warning printed that benchmarks without `--release` are not useful.