@r00ster91 2018-10-12
Yes I'm running it with `--release` and I don't get any warning.
Even when I'm running it 5 times in a row, "old fore" is always around `1.28×` to `1.36×` faster.
```
r00ster$ crystal test.cr --release
new fore 255.95M (  3.91ns) (±13.57%)  0 B/op        fastest
old fore 200.23M (  4.99ns) (± 9.60%)  0 B/op   1.28× slower
r00ster$ crystal test.cr --release
new fore 257.23M (  3.89ns) (±12.58%)  0 B/op        fastest
old fore 196.13M (   5.1ns) (±12.58%)  0 B/op   1.31× slower
r00ster$ crystal test.cr --release
new fore 260.85M (  3.83ns) (± 9.95%)  0 B/op        fastest
old fore 196.03M (   5.1ns) (±13.13%)  0 B/op   1.33× slower
r00ster$ crystal test.cr --release
new fore 259.37M (  3.86ns) (±10.97%)  0 B/op        fastest
old fore 197.99M (  5.05ns) (±12.43%)  0 B/op   1.31× slower
r00ster$ crystal test.cr --release
new fore 259.08M (  3.86ns) (±11.23%)  0 B/op        fastest
old fore 197.28M (  5.07ns) (±12.95%)  0 B/op   1.31× slower
```
...but when I now put the benchmark code in for example a `50.times` blocks, I get the same result as you:
```
new fore   12.4M ( 80.63ns) (± 9.38%)  0 B/op   1.00× slower
old fore  12.44M ( 80.37ns) (± 8.39%)  0 B/op        fastest
```
That's a bit strange...
But I guess you are right, the perfomance is the same.